Add files via upload
Analyzing a dataset of a **Ecommerce platform**. The Objective of this project is to build a model via **Supervised Learning** capable of predicting the **Expected Yearly Expenditure** of customers based on their **Presence and Non-Purchase Past behaviors** on the platform. 

We'll be conducting an **Exploratory Data Analysis** to understand the impact of each variable. And finally picking and building the most suitable model [**Linear Regression** for this project] for the use-case.